REFACTO: using Redux TK new configureStore method
