+++ src/reducers/root.js
@@ -1,19 +1,19 @@
 /**
  * Combine all reducers in this file and export the combined reducers.
  */
-
 import { combineReducers } from 'redux';
 import counterReducer from './clickCounter';
 
 /**
  * Merges the main reducer with the router state and dynamically injected reducers
  */
-export default function createReducer(injectedReducers = {}) {
+function createRootReducer(injectedReducers = {}) {
   const rootReducer = combineReducers({
-    // language: languageProviderReducer,
     counter: counterReducer,
     ...injectedReducers,
   });
 
   return rootReducer;
 }
+
+export default createRootReducer;
